**2022/09/04** * 重新绘制了设置碎片使之符合华为墨水屏设备风格

+++ app/src/main/java/com/jack/bookshelf/help/storage/Restore.kt
@@ -112,7 +112,7 @@ object Restore {
                     is String -> edit.putString(it.key, value)
                     else -> Unit
                 }
-                edit.putInt("versionCode", MApplication.getVersionCode())
+                edit.putLong("versionCode", MApplication.getVersionCode())
                 edit.apply()
             }
             e.onSuccess(true)
